# apps/api/swag/openapi3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vyzorix Update Server API",
    "version": "0.0.01",
    "description": "Go-generated OpenAPI contract from handler annotations"
  },
  "paths": {
...
    "/telemetry/{imei}": {
      "get": {
        "summary": "",
        "description": "",
        "tags": [
          "telemetry"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}